Add ESLint and TypeScript
Added ESLint and TypeScript to the project so that things can be nice and
clean 'n stuff. Hopefully we end up with better code by using these tools.
Had to add the line that exports all css as modules so that the TypeScript
doesn't get angry in next-env.d.ts. Also added the Editor component and
style sheet from the tutorial.