Show Errors Nicely
If we get errors, showing them in a more friendly way that pops up from
the bottom and doesn't kill the gems already on the screen. This makes
the page way less twitchy and less scary when you're writing up the code,
so should make users more comfortable to have errors while they're putting
stuff in, hopefully.