Hide Text Outside Gem
Hiding the text outside of the gems so that when they  got a lot of stuff
in 'em it looks a bit better on the screen. Making it so that they can
see all of it if they hover/touch the gem. It's not perfect, but it'll
do for now...hmmm....