Format Functions
Making it so that the functions display a bit prettier, like getting rid
of the unescaped newlines that might confuse someone, and then left aligning
the text so they're more natural to look at. They've got nasty overlay
with the other gems, but we can probably fix that later. For now, this'll
work for MVP. Also getting rid of the surrounding quotes for Symbols and
BigInts since they're not strings and we don't want to confuse people if
we don't have to.